Remove overwrites from example extension and embed into actual sheet functionality
